[REFACTOR/#79] 파일 정리 및 onboardingviewmodel 정리
