+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.splash.SplashActivity"
+            android:name=".presentation.onboarding.activity.OnBoardingActivity"
             android:exported="true"
             android:screenOrientation="portrait"
             android:theme="@style/Theme.YELLO.Splash">
@@ -44,11 +44,7 @@
                     android:scheme="kakao${NATIVE_APP_KEY}" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".presentation.onboarding.activity.OnBoardingActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"
-            android:windowSoftInputMode="adjustNothing" />
+
         <activity
             android:name=".presentation.auth.SignInActivity"
             android:exported="false"
