+++ app/src/main/java/com/el/yello/presentation/onboarding/activity/OnBoardingViewModel.kt
@@ -28,97 +28,29 @@ class OnBoardingViewModel @Inject constructor(
     private val onboardingRepository: OnboardingRepository,
     private val authRepository: AuthRepository,
 ) : ViewModel() {
-    private val _postSignupState = MutableLiveData<UiState<UserInfo>>()
-    val postSignupState: LiveData<UiState<UserInfo>>
-        get() = _postSignupState
-
-    private val _getValidYelloId = MutableLiveData<UiState<Boolean>>()
-    val getValidYelloId: LiveData<UiState<Boolean>>
-        get() = _getValidYelloId
-
-    private val _schoolData = MutableLiveData<UiState<SchoolList>>()
-    val schoolData: MutableLiveData<UiState<SchoolList>> = _schoolData
-
-    private val _departmentData = MutableLiveData<UiState<GroupList>>()
-    val departmentData: MutableLiveData<UiState<GroupList>> = _departmentData
-
-    private val _currentPage = MutableLiveData(0)
-    val currentPage: LiveData<Int> = _currentPage
-
-    var schoolPage = -1
-    private var isSchoolPagingFinish = false
-    private var totalSchoolPage = Integer.MAX_VALUE
-
-    var departmentPage = -1
-    private var isDepartmentPagingFinish = false
-    private var totalDepartmentPage = Integer.MAX_VALUE
-
-    val _school = MutableLiveData("")
-    val school: String
-        get() = _school.value?.trim() ?: ""
-
-    var kakaoId: String = ""
-    var email: String = ""
-    var profileImg: String = ""
-
-    private val _groupId = MutableLiveData<Long>()
-    val groupId: Long
-        get() = requireNotNull(_groupId.value)
-
-    val _department = MutableLiveData("")
-    private val department: String
-        get() = _department.value?.trim() ?: ""
-
-    val _studentId = MutableLiveData<Int>()
-    private val studentId: Int
-        get() = requireNotNull(_studentId.value)
-
-    val _name = MutableLiveData("")
-    private val name: String
-        get() = _name.value?.trim() ?: ""
-
-    val _id = MutableLiveData("")
-    private val id: String
-        get() = _id.value?.trim() ?: ""
-
-    val _gender = MutableLiveData("")
-    val gender: String
-        get() = _gender.value ?: ""
-
-    val _code = MutableLiveData("")
-    val code: String
-        get() = _code.value?.trim() ?: ""
-
-    private val _profile = MutableLiveData("")
-    val profile: String
-        get() = _profile.value ?: ""
-
-    private val _recommendId = MutableLiveData("")
-    val recommendId: String
-        get() = _recommendId.value ?: ""
-
-    val isValidSchool: LiveData<Boolean> = _school.map { school -> checkValidSchool(school) }
-    val isEmptyDepartment: LiveData<Boolean> =
-        _department.map { department -> this.checkEmpty(department) }
-
-    val isValidName: LiveData<Boolean> = _name.map { name -> checkName(name) }
-    val isValidId: LiveData<Boolean> = _id.map { id -> checkId(id) }
-
-    val isEmptyCode: LiveData<Boolean> = _code.map { code -> checkEmpty(code) }
-
-    private val _studentIdResult: MutableLiveData<List<Int>> = MutableLiveData()
-    val studentIdResult: LiveData<List<Int>> = _studentIdResult
-
-    private val _friendState = MutableLiveData<FriendList>()
-    val friendState: LiveData<FriendList> = _friendState
-
-    var kakaoFriendList: List<String> = listOf()
-    var selectedFriendIdList: List<Long> = listOf()
-    var selectedFriendCount: MutableLiveData<Int> = MutableLiveData(0)
 
     private var currentFriendPage: Int = -1
     private var isFriendPagingFinish = false
     private var totalFriendPage = Int.MAX_VALUE
+
+    // 학교 viewmodel (step 1)
+    private val _schoolData = MutableLiveData<UiState<SchoolList>>()
+    val _school = MutableLiveData("")
+    val schoolData: MutableLiveData<UiState<SchoolList>> = _schoolData
+    val school: String
+        get() = _school.value?.trim() ?: ""
+    val isValidSchool: LiveData<Boolean> = _school.map { school -> checkValidSchool(school) }
+    private fun checkValidSchool(school: String): Boolean {
+        return school.isNotBlank()
+    }
+
+    fun setSchool(school: String) {
+        _school.value = school
+    }
+
+    fun clearSchoolData() {
+        _schoolData.value = UiState.Success(SchoolList(0, emptyList()))
+    }
 
     // TODO: throttle 및 페이징 처리
     fun getSchoolList(search: String) {
@@ -149,6 +81,45 @@ class OnBoardingViewModel @Inject constructor(
                 }
             }
         }
+    }
+
+    // 학과 학번 viewmodel (step 2)
+    val _department = MutableLiveData("")
+    private val department: String get() = _department.value?.trim() ?: ""
+
+    private val _departmentData = MutableLiveData<UiState<GroupList>>()
+    val departmentData: MutableLiveData<UiState<GroupList>> = _departmentData
+
+    private fun checkEmpty(input: String) = input.isBlank()
+    val isEmptyDepartment: LiveData<Boolean> =
+        _department.map { department -> this.checkEmpty(department) }
+
+    private val _groupId = MutableLiveData<Long>()
+    val _studentId = MutableLiveData<Int>()
+
+    private val _studentIdResult: MutableLiveData<List<Int>> = MutableLiveData()
+    val studentIdResult: LiveData<List<Int>> = _studentIdResult
+    val groupId: Long
+        get() = requireNotNull(_groupId.value)
+    private val studentId: Int
+        get() = requireNotNull(_studentId.value)
+
+    fun setGroupInfo(department: String, groupId: Long) {
+        _department.value = department
+        _groupId.value = groupId
+    }
+
+    fun setStudentId(studentId: Int) {
+        _studentId.value = studentId
+    }
+
+    fun addStudentId() {
+        val mockList = listOf(15, 16, 17, 18, 19, 20, 21, 22, 23)
+        _studentIdResult.value = mockList
+    }
+
+    fun clearDepartmentData() {
+        _departmentData.value = UiState.Success(GroupList(0, emptyList()))
     }
 
     // TODO: throttle 및 페이징 처리
@@ -183,6 +154,45 @@ class OnBoardingViewModel @Inject constructor(
             }
         }
     }
+    // 이름 아이디 viewmodel (step 3)
+
+    val _name = MutableLiveData("")
+    val _id = MutableLiveData("")
+    private val name: String
+        get() = _name.value?.trim() ?: ""
+    private val id: String
+        get() = _id.value?.trim() ?: ""
+
+    val isValidName: LiveData<Boolean> = _name.map { name -> checkName(name) }
+    val isValidId: LiveData<Boolean> = _id.map { id -> checkId(id) }
+    private fun checkName(name: String) = Pattern.matches(REGEX_NAME_PATTERN, name)
+    private fun checkId(id: String) = Pattern.matches(REGEX_ID_PATTERN, id)
+
+    // 성별 viewmodel (step 4)
+
+    val _gender = MutableLiveData("")
+    val gender: String
+        get() = _gender.value ?: ""
+
+    fun selectGender(gender: String) {
+        _gender.value = gender
+    }
+    // 친구 추가 viewmodel (step 5)
+
+    var kakaoId: String = ""
+    var email: String = ""
+    var profileImg: String = ""
+
+    private val _profile = MutableLiveData("")
+    val profile: String
+        get() = _profile.value ?: ""
+
+    private val _friendState = MutableLiveData<FriendList>()
+    val friendState: LiveData<FriendList> = _friendState
+
+    var kakaoFriendList: List<String> = listOf()
+    var selectedFriendIdList: List<Long> = listOf()
+    var selectedFriendCount: MutableLiveData<Int> = MutableLiveData(0)
 
     fun addFriendList(friendGroup: FriendGroup) {
         viewModelScope.launch {
@@ -201,6 +211,51 @@ class OnBoardingViewModel @Inject constructor(
                 Timber.e(it.message)
             }
         }
+    }
+
+    // 추천인 코드 viewmodel (step 6)
+
+    val _code = MutableLiveData("")
+    val code: String
+        get() = _code.value?.trim() ?: ""
+
+    private val _recommendId = MutableLiveData("")
+    val recommendId: String
+        get() = _recommendId.value ?: ""
+
+    fun getValidYelloId() {
+        viewModelScope.launch {
+            onboardingRepository.getValidYelloId(yelloId = id)
+                .onSuccess { isValid ->
+                    Timber.d("GET VALID YELLO ID SUCCESS : $isValid")
+                    if (isValid == null) {
+                        _getValidYelloId.value = UiState.Empty
+                        return@launch
+                    }
+
+                    _getValidYelloId.value = UiState.Success(isValid)
+                }
+                .onFailure { t ->
+                    if (t is HttpException) {
+                        Timber.e("GET VALID YELLO ID FAILURE : $t")
+                        _getValidYelloId.value = UiState.Failure(t.code().toString())
+                        return@launch
+                    }
+                    Timber.e("GET VALID YELLO ID ERROR : $t")
+                }
+        }
+    }
+
+    // 앱 시작 회원가입 viewmodel (step 7)
+
+    private val _postSignupState = MutableLiveData<UiState<UserInfo>>()
+    val postSignupState: LiveData<UiState<UserInfo>> get() = _postSignupState
+
+    private val _getValidYelloId = MutableLiveData<UiState<Boolean>>()
+    val getValidYelloId: LiveData<UiState<Boolean>> get() = _getValidYelloId
+
+    fun validYellIdLoading() {
+        _getValidYelloId.value = UiState.Loading
     }
 
     fun postSignup() {
@@ -239,79 +294,15 @@ class OnBoardingViewModel @Inject constructor(
         }
     }
 
-    fun getValidYelloId() {
-        viewModelScope.launch {
-            onboardingRepository.getValidYelloId(yelloId = id)
-                .onSuccess { isValid ->
-                    Timber.d("GET VALID YELLO ID SUCCESS : $isValid")
-                    if (isValid == null) {
-                        _getValidYelloId.value = UiState.Empty
-                        return@launch
-                    }
-
-                    _getValidYelloId.value = UiState.Success(isValid)
-                }
-                .onFailure { t ->
-                    if (t is HttpException) {
-                        Timber.e("GET VALID YELLO ID FAILURE : $t")
-                        _getValidYelloId.value = UiState.Failure(t.code().toString())
-                        return@launch
-                    }
-                    Timber.e("GET VALID YELLO ID ERROR : $t")
-                }
-        }
-    }
-
-    fun setSchool(school: String) {
-        _school.value = school
-    }
-
-    fun setGroupInfo(department: String, groupId: Long) {
-        _department.value = department
-        _groupId.value = groupId
-    }
-
-    fun setStudentId(studentId: Int) {
-        _studentId.value = studentId
-    }
-
-    fun selectGender(gender: String) {
-        _gender.value = gender
-    }
-
-    fun clearSchoolData() {
-        _schoolData.value = UiState.Success(SchoolList(0, emptyList()))
-    }
-
-    fun clearDepartmentData() {
-        _departmentData.value = UiState.Success(GroupList(0, emptyList()))
-    }
-
-    private fun checkValidSchool(school: String): Boolean {
-        return school.isNotBlank()
-    }
-
-    private fun checkEmpty(input: String) = input.isBlank()
-
-    private fun checkName(name: String) = Pattern.matches(REGEX_NAME_PATTERN, name)
-
-    private fun checkId(id: String) = Pattern.matches(REGEX_ID_PATTERN, id)
-
+    // 화면 이동
+    private val _currentPage = MutableLiveData(0)
+    val currentPage: LiveData<Int> = _currentPage
     fun navigateToNextPage() {
         _currentPage.value = currentPage.value?.plus(1)
     }
 
     fun navigateToBackPage() {
         _currentPage.value = currentPage.value?.minus(1)
-    }
-
-    fun addStudentId() {
-        val mockList = listOf(15, 16, 17, 18, 19, 20, 21, 22, 23)
-        _studentIdResult.value = mockList
-    }
-
-    fun validYellIdLoading() {
-        _getValidYelloId.value = UiState.Loading
     }
 
     companion object {
