+++ app/src/main/java/com/el/yello/presentation/onboarding/fragment/school/SchoolFragment.kt
@@ -8,7 +8,7 @@ import com.el.yello.R
 import com.el.yello.databinding.FragmentSchoolBinding
 import com.el.yello.presentation.auth.SocialSyncActivity
 import com.el.yello.presentation.onboarding.activity.OnBoardingViewModel
-import com.el.yello.presentation.onboarding.fragment.school.university.SearchDialogSchoolFragment
+import com.el.yello.presentation.onboarding.fragment.school.dialog.SearchDialogSchoolFragment
 import com.example.ui.base.BindingFragment
 import com.example.ui.view.setOnSingleClickListener
 
@@ -22,10 +22,6 @@ class SchoolFragment : BindingFragment<FragmentSchoolBinding>(R.layout.fragment_
         initSearchSchoolBtnClickListener()
         setConfirmBtnClickListener()
         setupSchool()
-        binding.btnSchoolBackBtn.setOnSingleClickListener {
-            val intent = Intent(getActivity(), SocialSyncActivity::class.java)
-            startActivity(intent)
-        }
     }
 
     private fun initSearchSchoolBtnClickListener() {
@@ -37,6 +33,10 @@ class SchoolFragment : BindingFragment<FragmentSchoolBinding>(R.layout.fragment_
     private fun setConfirmBtnClickListener() {
         binding.btnSchoolNext.setOnSingleClickListener {
             viewModel.navigateToNextPage()
+        }
+        binding.btnSchoolBackBtn.setOnSingleClickListener {
+            val intent = Intent(activity, SocialSyncActivity::class.java)
+            startActivity(intent)
         }
     }
